+++ example/lib/timer_viewmodel.dart
@@ -4,17 +4,19 @@ import 'ticker.dart';
 
 class TimerViewModel extends ChangeNotifier {
   TimerViewModel({required Ticker ticker}) : _ticker = ticker {
-    _initCommand();
+    _initCommands();
   }
 
   static const int _duration = 30;
   final Ticker _ticker;
 
   late StreamCommand<int, NoParams> timerCommand;
+  late Command<void, NoParams> startCommand;
+  late Command<void, NoParams> stopCommand;
 
   int timeLeft = _duration;
 
-  void _initCommand() {
+  void _initCommands() {
     timerCommand = StreamCommand<int, NoParams>((_) {
       return _ticker
           .tick(ticks: _duration)
@@ -22,19 +24,23 @@ class TimerViewModel extends ChangeNotifier {
     });
 
     timerCommand.latestResult.addListener(_onResultChanged);
+    startCommand = Command<void, NoParams>(_start);
+    stopCommand = Command<void, NoParams>(_stop);
   }
 
-  void start() {
+  Future<Result<void>> _start(NoParams _) async {
     if (timerCommand.isListening.value) {
       timerCommand.stop();
     }
     timerCommand.start(const NoParams());
+    return const Success(null);
   }
 
-  void stop() {
+  Future<Result<void>> _stop(NoParams _) async {
     timerCommand.stop();
     timeLeft = _duration;
     notifyListeners();
+    return const Success(null);
   }
 
   void _onResultChanged() {